requirements.txt
django
djangorestframework
django-cors-headers 
djangorestframework-simplejwt
PyJWT
python-dotenv